move querying of course_aggregates_table to server side
